requirements.txt
attrs==22.1.0
bcrypt==4.0.0
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
neo4j-driver==5.0.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyJWT==2.5.0
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
pytz==2022.2.1
six==1.16.0
tomli==2.0.1
Werkzeug==2.2.2

neo4j~=5.16.0
setuptools~=68.2.0